update: lambda deploy
